feat(hocuspocus): extract into standalone Docker service
- Move Hocuspocus from apps/docs into services/hocuspocus/ workspace
- Refactor persistence and auth with dependency injection (no singletons)
- Add HTTP health check endpoint on port 1241
- Create multi-stage Dockerfile (no native deps needed)
- Update apps/docs to pure static server (remove Hocuspocus startup)
- Simplify docs Dockerfile (remove API source compilation, native deps)
- Update apps/api to consume @gruenerator/hocuspocus workspace package
- Add hocuspocus service to docker-compose.{prod,test}.yml
- Route /ws upstream from docs:1240 to hocuspocus:1240 in nginx
- Add build-hocuspocus job to CI build-images workflow
- Delete apps/api/services/hocuspocus/ (moved to services/hocuspocus/)
