+++ .github/workflows/build-images.yml
@@ -45,6 +45,11 @@ on:
         required: false
         default: 'false'
         type: boolean
+      build_hocuspocus:
+        description: 'Build hocuspocus image only'
+        required: false
+        default: 'false'
+        type: boolean
       build_chat:
         description: 'Build chat image only'
         required: false
@@ -62,6 +67,7 @@ jobs:
       web: ${{ steps.filter.outputs.web }}
       api: ${{ steps.filter.outputs.api }}
       docs: ${{ steps.filter.outputs.docs }}
+      hocuspocus: ${{ steps.filter.outputs.hocuspocus }}
       mcp: ${{ steps.filter.outputs.mcp }}
       doku: ${{ steps.filter.outputs.doku }}
       chat: ${{ steps.filter.outputs.chat }}
@@ -90,6 +96,8 @@ jobs:
               - 'apps/docs/**'
               - 'packages/shared/**'
               - 'packages/docs/**'
+            hocuspocus:
+              - 'services/hocuspocus/**'
             mcp:
               - 'services/mcp/**'
               - 'packages/shared/**'
@@ -235,6 +243,50 @@ jobs:
           cache-from: type=gha
           cache-to: type=gha,mode=max
 
+  build-hocuspocus:
+    needs: detect-changes
+    if: >
+      needs.detect-changes.outputs.hocuspocus == 'true' ||
+      github.event.inputs.force_all == 'true' ||
+      github.event.inputs.build_hocuspocus == 'true'
+    runs-on: ubuntu-latest
+    permissions:
+      contents: read
+      packages: write
+    steps:
+      - uses: actions/checkout@v4
+
+      - name: Set up Docker Buildx
+        uses: docker/setup-buildx-action@v3
+
+      - name: Log in to Container Registry
+        uses: docker/login-action@v3
+        with:
+          registry: ${{ env.REGISTRY }}
+          username: ${{ github.actor }}
+          password: ${{ secrets.GITHUB_TOKEN }}
+
+      - name: Extract metadata
+        id: meta
+        uses: docker/metadata-action@v5
+        with:
+          images: ${{ env.IMAGE_PREFIX }}-hocuspocus
+          tags: |
+            type=ref,event=branch
+            type=sha,prefix=
+            type=raw,value=latest,enable={{is_default_branch}}
+
+      - name: Build and push
+        uses: docker/build-push-action@v5
+        with:
+          context: .
+          file: services/hocuspocus/Dockerfile
+          push: true
+          tags: ${{ steps.meta.outputs.tags }}
+          labels: ${{ steps.meta.outputs.labels }}
+          cache-from: type=gha
+          cache-to: type=gha,mode=max
+
   build-mcp:
     needs: detect-changes
     if: >
@@ -370,7 +422,7 @@ jobs:
           cache-to: type=gha,mode=max
 
   summary:
-    needs: [build-web, build-api, build-docs, build-mcp, build-doku, build-chat]
+    needs: [build-web, build-api, build-docs, build-hocuspocus, build-mcp, build-doku, build-chat]
     if: always()
     runs-on: ubuntu-latest
     permissions:
@@ -384,6 +436,7 @@ jobs:
           echo "| web | ${{ needs.build-web.result }} |" >> $GITHUB_STEP_SUMMARY
           echo "| api | ${{ needs.build-api.result }} |" >> $GITHUB_STEP_SUMMARY
           echo "| docs | ${{ needs.build-docs.result }} |" >> $GITHUB_STEP_SUMMARY
+          echo "| hocuspocus | ${{ needs.build-hocuspocus.result }} |" >> $GITHUB_STEP_SUMMARY
           echo "| mcp | ${{ needs.build-mcp.result }} |" >> $GITHUB_STEP_SUMMARY
           echo "| doku | ${{ needs.build-doku.result }} |" >> $GITHUB_STEP_SUMMARY
           echo "| chat | ${{ needs.build-chat.result }} |" >> $GITHUB_STEP_SUMMARY
