+++ apps/api/package.json
@@ -20,7 +20,7 @@
     "setup:authentik-sources:force": "node scripts/setupAuthentikSources.js --force",
     "start:dev": "cross-env NODE_ENV=development npm run start:backend",
     "start:dev:no-yjs": "cross-env NODE_ENV=development YJS_ENABLED=false npm run start:backend",
-    "start:hocuspocus": "npx tsx services/hocuspocus/hocuspocusServer.ts",
+    "start:hocuspocus": "npx tsx ../../services/hocuspocus/src/index.ts",
     "start:prod": "cross-env NODE_ENV=production npm run start:backend",
     "test:embedding": "node scripts/testEmbedding.mjs",
     "download-whisper": "npx nodejs-whisper download",
@@ -48,9 +48,8 @@
     "@ai-sdk/openai": "^3.0.0",
     "@anthropic-ai/sdk": "^0.71.0",
     "@blocknote/xl-ai": "^0.46.2",
+    "@gruenerator/hocuspocus": "workspace:*",
     "@gruenerator/shared": "workspace:*",
-    "@hocuspocus/extension-logger": "^3.4.3",
-    "@hocuspocus/server": "^3.4.3",
     "@langchain/core": "^1.0.2",
     "@langchain/langgraph": "^1.0.1",
     "@mistralai/mistralai": "^1.7.2",
