+++ apps/docs/docker-compose.yml
@@ -9,7 +9,6 @@ services:
     restart: unless-stopped
     ports:
       - '3000:3000'
-      - '1240:1240'
     environment:
       # Node.js
       NODE_ENV: production
@@ -17,28 +16,12 @@ services:
       # Server Configuration
       PORT: 3000
 
-      # Hocuspocus WebSocket Server
-      HOCUSPOCUS_PORT: 1240
-      HOCUSPOCUS_HOST: 0.0.0.0
-
-      # Database (PostgreSQL)
-      # Replace these with your actual database credentials
-      POSTGRES_HOST: ${POSTGRES_HOST:-host.docker.internal}
-      POSTGRES_PORT: ${POSTGRES_PORT:-5432}
-      POSTGRES_DB: ${POSTGRES_DB:-gruenerator}
-      POSTGRES_USER: ${POSTGRES_USER:-postgres}
-      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-your_password}
-
       # API Base URL (for frontend API calls)
       VITE_API_BASE_URL: ${VITE_API_BASE_URL:-/api}
       VITE_API_TARGET: ${VITE_API_TARGET:-http://localhost:3001}
 
       # Hocuspocus WebSocket URL (for frontend WebSocket connections)
       VITE_HOCUSPOCUS_URL: ${VITE_HOCUSPOCUS_URL:-ws://localhost:1240}
-
-    # Optional: Connect to external PostgreSQL
-    # networks:
-    #   - gruenerator-network
 
     healthcheck:
       test:
@@ -53,7 +36,40 @@ services:
       retries: 3
       start_period: 40s
 
-# Optional: Define network for connecting to external services
-# networks:
-#   gruenerator-network:
-#     external: true
+  hocuspocus:
+    build:
+      context: ../..
+      dockerfile: services/hocuspocus/Dockerfile
+    container_name: gruenerator-hocuspocus
+    restart: unless-stopped
+    ports:
+      - '1240:1240'
+      - '1241:1241'
+    environment:
+      NODE_ENV: production
+      HOCUSPOCUS_PORT: 1240
+      HOCUSPOCUS_HOST: 0.0.0.0
+      HEALTH_PORT: 1241
+
+      # Database (PostgreSQL)
+      POSTGRES_HOST: ${POSTGRES_HOST:-host.docker.internal}
+      POSTGRES_PORT: ${POSTGRES_PORT:-5432}
+      POSTGRES_DB: ${POSTGRES_DB:-gruenerator}
+      POSTGRES_USER: ${POSTGRES_USER:-postgres}
+      POSTGRES_PASSWORD: ${POSTGRES_PASSWORD:-your_password}
+
+      # Redis
+      REDIS_URL: ${REDIS_URL:-redis://host.docker.internal:6379}
+
+      # Auth
+      SESSION_SECRET: ${SESSION_SECRET:-fallback-secret-please-change}
+
+    extra_hosts:
+      - 'host.docker.internal:host-gateway'
+
+    healthcheck:
+      test: ['CMD', 'curl', '-f', 'http://localhost:1241/health']
+      interval: 30s
+      timeout: 10s
+      retries: 3
+      start_period: 10s
