+++ apps/docs/package.json
@@ -22,9 +22,7 @@
     "@floating-ui/react": "^0.27.16",
     "@gruenerator/docs": "workspace:*",
     "@gruenerator/shared": "workspace:*",
-    "@hocuspocus/extension-logger": "^3.4.3",
     "@hocuspocus/provider": "^3.4.3",
-    "@hocuspocus/server": "^3.4.3",
     "@mantine/core": "^8.3.14",
     "@mantine/hooks": "^8.3.14",
     "@radix-ui/react-dropdown-menu": "^2.1.16",
@@ -58,13 +56,11 @@
     "express": "^5",
     "helmet": "^8.1.0",
     "lodash.throttle": "^4.1.1",
-    "pg": "^8.16.3",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
     "react-hotkeys-hook": "^5.2.1",
     "react-icons": "^5.5.0",
     "react-router-dom": "^7.12.0",
-    "winston": "^3.17.0",
     "y-websocket": "^2.0.4",
     "yjs": "^13.6.20",
     "zustand": "^5.0.9"
@@ -76,7 +72,6 @@
     "@types/express": "^5.0.6",
     "@types/lodash.throttle": "^4.1.9",
     "@types/node": "^24.10.4",
-    "@types/pg": "^8.16.0",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.5",
     "@vitejs/plugin-react": "^4.4.1",
