+++ nginx.conf
@@ -52,7 +52,7 @@ http {
 
         # Hocuspocus WebSocket — real-time collaboration
         location /ws {
-            set $upstream_hocuspocus http://docs:1240;
+            set $upstream_hocuspocus http://hocuspocus:1240;
             proxy_pass $upstream_hocuspocus;
             proxy_http_version 1.1;
             proxy_set_header Upgrade $http_upgrade;
