+++ pnpm-workspace.yaml
@@ -2,5 +2,6 @@ packages:
   - 'apps/*'
   - 'packages/*'
   - 'services/comfyui'
+  - 'services/hocuspocus'
   - 'services/mcp'
   - 'documentation'
