+++ services/hocuspocus/.env.example
@@ -1,0 +1,21 @@
+# Hocuspocus WebSocket Server
+HOCUSPOCUS_PORT=1240
+HOCUSPOCUS_HOST=0.0.0.0
+HEALTH_PORT=1241
+
+# PostgreSQL
+POSTGRES_HOST=localhost
+POSTGRES_PORT=5432
+POSTGRES_DB=gruenerator
+POSTGRES_USER=postgres
+POSTGRES_PASSWORD=your_password
+
+# Redis
+REDIS_URL=redis://localhost:6379
+
+# Auth
+SESSION_SECRET=your-session-secret
+
+# Logging
+LOG_LEVEL=info
+NODE_ENV=development
